Move rotated-rect-math utils
